Add files via upload
Whenever you find your device paused on an Eeros network you can use this script to bypass it. Be prepared if they have IDS though, because this will spam their inbox with notifications of new devices on the network. As with any tool, I am not responsible for any malicious use of this script. Do not use this script on networks that you don't have express permission to exploit.

USE AT YOUR OWN RISK!